requirements.txt
streamlit==1.28.0
transformers==4.35.0
torch==2.1.0
scipy==1.11.0
zhipuai==2.0.1
python-dotenv==1.0.0
numpy==1.24.0
soundfile==0.12.0
accelerate==0.24.0